GNPy: fix applying power config to devices
The code was reading data from the reverse path even for the a-z path,
and then the actual applying was reading from wrong variables.

Don't call the devices "ingress" and "egress". This isn't really wrong
when looking at a link, but I think it's confusing. The term "ingress"
and "egress" might as well refer to the input and output ports of a
device, and clearly one cannot "set the input port" (even though I often
with for that).

Change-Id: I8a848f4ae089e46ba20f73f13cab77ae23ffa915